Made some pretty big headway tonight WRT the task's code inclusion, task id counting, and building of a composite app, in an automated manner.  Still have some work to do on automating the address pointer pool polling (#def loop is the answer, with macro expansion on the id), and need to further the namespace for each included task source.
The new approach will do away with a need for a task prototype, I believe, as the user will be abstracted away from having to worry about that, with the automation that I am building into things.  Rrght now each task's main loop and all helper functions/subs are being actively renamed during preprocessing/compilation - to drive namespace unicity.  The next steps include -) automating the buildup of the task entry pointers, -) incorporating the same, or conjuring up a similar namespace approach for dims, constants, labels, and possibly the #defs that are in each task's source code and, of course, the context switching and other shite that goes with a proper multi-tasking OS.

The automated preprocessor task counting kicked my arse royally.  It's been something that I've been battling for the last umpteen days, in figuring out a way to make the damn thing count and keep track of the number of ad-hoc tasks being included, during preprocessing/compile-time.  I ended up resorting to some pretty cute preprocessor magic.  This is, however, going to drive an overt requirement for FilePP going forward.  There is simply no way around it, given my selected approach (abstracting the user away from having to trikfuk each task's code just to make it become an importable task).  The name space management for each task ended up requiring the use of the regex features of FilePP (more will be needed to do the dims/constants/labels (& #defs if I go therre)).  The naming and counting for each task ended up imputing a use of FilePP's math constructs (although I did find a different approach using only the #defplus featureset of FilePP - it is less elegant and no more cumbersome, so I went ahead and took a combined approach with using both #defplus and the maths module).

Coming back to the requirement for FilePP.  I have been, in the rear of muh brain housing assembly, mulling on a way to make use of BPP with this.  The only thing I can come up with is to have a two-phased preprocessing.  The first phase would be the automated namespace manipulation and task inclusion, and then passing off a partially processed bpp for BPP.exe to run on.  Frankly, that is just dickey, especially when FilePP can do it all in a single pass.  It would only add complexity and drive further changes to BT (like I haven't triked the frak out of it already...).  It boils down to time.  No sense expending same at this point...