Ok, was a bit premature early this morning, when I asserted that the name space issues were maturing. ...  They are now.  The catch is that in the user tasks, variables, constants, labels, functions, and subs must have a name that starts with 'ABT_'.  That way, when the preprocessor does it's magic, these ABT_blah are macro extanded to ABT_x_blah, where x is the task number.  If user is to do a goto to main, s/he must put a different label immediately following main: and avoid the use of the work main therein.  And, lastly, the user must concatenate his/her end sub (to endsub), and end function (to endfulction).  I still need to verify if the other instances of endBLAH to see if there is a requirement for special handling, or if a simple user concatenation is appropriate.   I suspect the latter will work, as I perceive that the regex I am using to change Main's end to endsub might fire on these other uses of an unconcatenated end (endselect, and endif).
The attached TasksConfig demonstrates instantiation of 24 tasks being loaded (12 instances of the blinkey/printy task pairs), just to demonstrate the inclusion process is as robust as expected at this point, and that my regex is firing/replacing as appropriate, and that my counting and looping routines are functional as expected.  Right now they appear to be.  I am pretty effin happy with where things are at ATM.  I am really getting down to the point where task oversight and context switching is the last big hurdles that I perceive will tax muh gray matter.  The queues (messages, mutexs, semaphores) should be pretty straight forward, once I get the tasks pushing and popping to the beat of the timer.  Ditto for the lists/tables, however I suspect that the lists/tables will have to mature concurrently with the context switching, so I can prioritize the tasks, and do a RMA or preemptive/roundrobin scheduling construct.  The latter will likely be first, just to get me a sense of satisfaction which will likely turn into the culmination of ABmt ... And then the next growth spurts these efforts will have thereafter will end up birthing the early iterations of ABrtos...

Something that has been consistently ringing in the back of my gray matter is the megaspeed (~1MBps?) GPIO bitbanged serial trace/debug streaming implemenation that will be needed, so that I can have tools on the host to monitor and diagnose the implementation of the OS, and the tasks themselves.  I did some preliminary testing with ASM wiggling a GPIO pin and was able to get the 100MHz development target device (one of Bruce's 54102s) to toggle at 16MHz (https://i.imgur.com/Zjliv3z.png) with some highly optimized (2-instruction) ASM.  I'm suspecting that I'll be lucky to get anywhere near 1MBps out.  Conserviatively I suspect it will be somewhere near or less than half that, which is still a whopping 5Mbps (hopefully) which is plenty fast to be able to dprint bitbanged debug data out a gpio during an ISR without fuking things up too badly (hopefully).

At any rate, I've droned on long enuf.  It is what it is.  (still chewing on BT upgrades too (integrating BT into TKE (as a plugin)) or taking portions of TKE and integrating them into BT to, eventually, have BT be a proper sourcecode editor, in addition to being the compilation/flashing toolchain - ah, grandiose plans I do not lack... :))

-t