Wow, chased my tail a bit the last hour - Frigging windoze update caused something to be remiss with aninstantiated instance of the serial port.
Awoke to a login screen, logged in and everything seemed normal.  I was albe to flash the target, but it would not start the app - acted like it did, but it would just hang prior to getting to main.

I thought it was my late hour coding last night.  Nope.  Ended up reconnecting the programmer and that resolved the issue, an hour after starting to scratch muh noggin.  This is why there is some commented-out #warning dialog littered about...

Also cleaned up some stuffs and added some explanatory dialog comments to some of the FilePP constructs I am using to implement the task counting and namespace approach.

TODAY I am starting to play with the actual context switch implementation.  Hopefully it won't be too painful.